fix issue related to cirpack bug.
cirpack is looking for an Expires header in REGISTER request.
If it doesn't find this expires header or any expires parameter
in contact, it sends back a 200 with expires=0 in contact header
and generates an exception in peers.

To avoid this issue, I added an Expires header in REGISTER request.